pyproject.toml
[tool.poetry]
name = "radio_db"
version = "0.1.8"
description = ""
authors = ["Jono Hill <jono@hillnz.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.7.4"
"ruamel.yaml" = "^0.17.4"
SQLAlchemy = "^1.4.13"
aiosqlite = "^0.17.0"
pydantic = "^1.8.1"
asyncpg = "^0.23.0"
spotipy = "^2.18.0"
python-dotenv = "^0.17.1"
typer = "^0.3.2"

[tool.poetry.dev-dependencies]
pylint = "^2.8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
